+++ bin/common.mli
@@ -8,7 +8,7 @@ val root : t -> Workspace_root.t
 val config : t -> Dune.Config.t
 val only_packages : t -> Dune.Package.Name.Set.t option
 val watch : t -> bool
-val default_target : t -> string
+val default_target : t -> Arg.Dep.t
 
 val prefix_target : t -> string -> string
 
@@ -16,12 +16,12 @@ val prefix_target : t -> string -> string
     [set_common_other common ~targets]. In general, [set_common] executes
     sequence of side-effecting actions to initialize Dune's working
     environment based on the options determined in a [Common.t] record *)
-val set_common : t -> targets:string list -> unit
+val set_common : t -> targets:Arg.Dep.t list -> unit
 
 (** [set_common_other common ~targets] sets all stateful values dictated by
     [common], except those accounted for by [set_dirs]. [targets] are
     used to obtain external library dependency hints, if needed. *)
-val set_common_other : t -> targets:string list -> unit
+val set_common_other : t -> targets:Arg.Dep.t list -> unit
 
 (** [set_dirs common] sets the workspace root and build directories, and makes
  the root the current working directory *)