+++ src/dune_file.ml
@@ -271,7 +271,7 @@ module Dep_conf = struct
       List [ Dune_lang.unsafe_atom_of_string "glob_files"
            ; String_with_vars.encode t ]
     | Source_tree t ->
-      List [ Dune_lang.unsafe_atom_of_string "files_recursively_in"
+      List [ Dune_lang.unsafe_atom_of_string "source_tree"
            ; String_with_vars.encode t ]
     | Package t ->
       List [ Dune_lang.unsafe_atom_of_string "package"
@@ -2319,7 +2319,7 @@ module Stanzas = struct
             loc.Loc.start.pos_lnum
         in
         User_error.raise ~loc
-          [ Pp.text "Recursive inclusion of jbuild files detected:"
+          [ Pp.text "Recursive inclusion of dune files detected:"
           ; Pp.textf "File %s is included from %s"
               (Path.Source.to_string_maybe_quoted file)
               (line_loc last)