+++ src/file_tree.mli
@@ -109,9 +109,3 @@ val dir_is_vendored : t -> Path.Source.t -> bool option
 
 (** [true] iff the path is a file *)
 val file_exists : t -> Path.Source.t -> bool
-
-val files_recursively_in
-  :  t
-  -> prefix_with:Path.t
-  -> Path.Source.t
-  -> Path.Set.t