+++ CHANGES.md
@@ -65,7 +65,7 @@
   and the dep spec `sandbox` in dune language. These let the user control the level of
   sandboxing done by dune per rule and globally. The rule specification takes precedence.
   The global configuration merely specifies the default.
-  (#2213, @aalekseyev, @jdimino)
+  (#2213, @aalekseyev, @diml)
 
 - Remove support for old style subsystems. Dune will now emit a warning to
   reinstall the library with the old style subsystem. (#2480, @rgrinberg)
@@ -78,6 +78,13 @@
 
 - Set up formatting rules by default. They can be configured through a new
   `(formatting)` stanza in `dune-project` (#2347, fixes #2315, @emillon)
+
+- Change default target from `@install` to `@all`. (#2449, fixes #1220,
+  @rgrinberg)
+
+- Cleanup stale directories when using `(source_tree ...)` in the
+  presence of directories with only sub-directories and no files
+  (#2514, fixes #2499, @diml)
 
 1.11.0 (23/07/2019)
 -------------------